Add Tasks

+++ src/main/java/hexlet/code/app/controller/TaskStatusController.java
@@ -1,10 +1,13 @@
 package hexlet.code.app.controller;
 
 import hexlet.code.app.dto.TaskStatusDto;
+import hexlet.code.app.model.Task;
 import hexlet.code.app.model.TaskStatus;
+import hexlet.code.app.repository.TaskRepository;
 import hexlet.code.app.repository.TaskStatusRepository;
-import hexlet.code.app.service.TaskStatusServiceImpl;
+import hexlet.code.app.service.TaskStatusService;
 import lombok.AllArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,6 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 import static hexlet.code.app.controller.TaskStatusController.TASK_STATUS_CONTROLLER_PATH;
 
@@ -31,7 +35,9 @@ public class TaskStatusController {
 
     private final TaskStatusRepository statusRepository;
 
-    private final TaskStatusServiceImpl statusService;
+    private final TaskStatusService statusService;
+
+    private final TaskRepository taskRepository;
 
     @GetMapping
     public List<TaskStatus> getAll() {
@@ -58,6 +64,12 @@ public class TaskStatusController {
 
     @DeleteMapping(path = "/{id}")
     public void deleteStatus(@PathVariable Long id) {
+        Optional<Task> taskWithThisStatus = taskRepository.findFirst1ByTaskStatusId(id);
+
+        if (taskWithThisStatus.isPresent()) {
+            throw new DataIntegrityViolationException("Unable to delete the status associated with an existing task");
+        }
+
         statusRepository.deleteById(id);
     }
 }
